stamina manage by allies

+++ Project/Entities/EntityCapacities.cs
@@ -2,7 +2,7 @@
 {
     public string _name { get; set; }
     public float _damage { get; set; }
-    public int _stamina { get; set; }
+    public float _stamina { get; set; }
     public int _speed { get; set; }
     public float _resistance { get; set; }
     public float _boostDamage { get; set; }
